feat: brightness from beam dwell, and stop capping vector detail at the field
Three related changes to what the vector renderers are given and what they do
with it.

Brightness now comes from dwell. Light per unit length is beam current times how
long the beam spent there, and the intensity code only ever supplied the current.
The anchor is derived rather than fitted: the beam's top speed is
0x1FF * 255 >> 4, which is 8.05 cycles per unit along an axis and 1/sqrt(2) of
that on a diagonal. Measured over Tempest, Quantum and Star Wars, the observed
floor is 5.7 cycles per unit and the 5th percentile is 8.0, which are those two
figures: the same deflection hardware in all three. A full-intensity vector at
that speed reads full white and anything slower blooms.

Display-list coordinates keep their fraction. A unit has no size of its own: it
is 1/65536 of the AVG's position accumulator, so how many units a picture spans
is decided by whatever scale values a game's programmers picked. Rounding
positions to whole units therefore charged Star Wars, whose data spans 320,
twice the precision penalty of Tempest, whose data spans 580, on the same
physical tube. VectorLine now carries f32 and the generator's 16 bits below the
unit survive.

And the two jobs display_size was doing are now separate. It was both the
coordinate extent the generator centres itself in and the pixel count to
rasterize into, so a machine's rendered detail was capped at whatever numeric
range its data used, and enlarging the buffer alone could not help: xcenter would
move with it and the picture would shrink into the middle. Renderable gains
vector_field_size for the extent, display_size becomes the raster resolution, and
the rasterizer maps one onto the other. Star Wars goes from 320x330 to 705x727
and is legible.

The raster resolution has a floor rather than a chosen number. The spot's size in
output pixels depends only on how many pixels the long axis has, the generator's
extent cancelling out, so setting that equal to MIN_SIGMA_PIXELS gives about 727
pixels: below it the beam has to be drawn wider than the tube's and detail the
generator computed is lost. Above it there is more to gain, which is a cost
decision for the display settings work.

Four tests on the dwell model: twice the dwell is twice the light, the beam's top
speed is where full white sits, a slower vector saturates, and a generator that
reports no travel time falls back to the intensity code rather than going black.
The first needed a dim vector to test at all, since at full intensity the quick
one already peaks and doubling it only clips.

Star Wars confirmed by eye. Part of phosphor-emulator-ukgy.3.

+++ frontend/src/vector_gl.rs
@@ -356,8 +356,8 @@ impl VectorRenderer {
                 brightness * (line.b as f32 / 255.0),
             );
 
-            let (x0, y0) = (line.x0 as f32, line.y0 as f32);
-            let (x1, y1) = (line.x1 as f32, line.y1 as f32);
+            let (x0, y0) = (line.x0, line.y0);
+            let (x1, y1) = (line.x1, line.y1);
             let (dx, dy) = (x1 - x0, y1 - y0);
             let len = (dx * dx + dy * dy).sqrt();
             // A zero-length vector is a dot: the beam arrived and did not
